Function spirv_reflect::ffi::spvReflectChangeOutputVariableLocation

pub unsafe extern "C" fn spvReflectChangeOutputVariableLocation(
    p_module: *mut SpvReflectShaderModule, 
    p_output_variable: *const SpvReflectInterfaceVariable, 
    new_location: u32
) -> SpvReflectResult

@fn spvReflectChangeOutputVariableLocation @brief Assign a new location to an output interface variable. In addition to updating the reflection data, this function modifies the underlying SPIR-V bytecode. The updated code can be retrieved with spvReflectGetCode(). It is the caller's responsibility to avoid assigning the same location to multiple output variables. If the output variable is used by multiple entry points in the module, it will be changed in all of them. @param p_module Pointer to an instance of SpvReflectShaderModule. @param p_output_variable Pointer to the output variable to update. @param new_location The new location to assign to p_output_variable. @return If successful, returns SPV_REFLECT_RESULT_SUCCESS. Otherwise, the error code indicates the cause of the failure.