Function spirv_reflect::ffi::spvReflectEnumerateEntryPointOutputVariables

pub unsafe extern "C" fn spvReflectEnumerateEntryPointOutputVariables(
    p_module: *const SpvReflectShaderModule, 
    entry_point: *const c_char, 
    p_count: *mut u32, 
    pp_variables: *mut *mut SpvReflectInterfaceVariable
) -> SpvReflectResult

@fn spvReflectEnumerateEntryPointOutputVariables @brief Enumerate the output variables for a given entry point. @param p_module Pointer to an instance of SpvReflectShaderModule. @param entry_point The name of the entry point to get the output variables for. @param p_count If pp_variables is NULL, the entry point's output variable count will be stored here. If pp_variables is not NULL, *p_count must contain the entry point's output variable count. @param pp_variables If NULL, the entry point's output variable count will be written to *p_count. If non-NULL, pp_variables must point to an array with p_count entries, where pointers to the entry point's output variables will be written. The caller must not free the interface variables written to this array. @return If successful, returns SPV_REFLECT_RESULT_SUCCESS. Otherwise, the error code indicates the cause of the failure.